normalize paths resolved via import map (#1)
This fixes a bug where the same module could be included twice in the
output bundle: When the module was imported with a relative specifier
(./mod.js) and with a bare specifier (stuff/mod.js), the full module
path would be resolved by esbuild in the first case and by the import
map plugin in the second. However, the import map plugin would return a
semantically equivalent, but different, path string (e.g.
/src/stuff/mod.js vs. /src/./stuff/mod.js) which apparently made esbuild
process it as two distinct modules.